requirements.txt
boto3
psycopg2-binary
pandas 
apache-airflow 

